Issue #11 : Create Netatmo class
