+++ .env.default
@@ -1,6 +1,7 @@
+APP_URL="APP_URL"
+
 # Netatmo API
 NETATMO_CLIENT_ID="CLIENT_ID"
 NETATMO_CLIENT_SECRET="CLIENT_SECRET"
 NETATMO_USERNAME="USERNAME"
 NETATMO_PASSWORD="PASSWORD"
-NETATMO_WEBHOOK="WEBHOOK"
